Fixed issues with calculating measure
The intersectionPoint often times was not intersecting with the nhdFlowline (even though it was found by spliting the downstreamPath with the nhdFlowline), so the nhdFlowline could not be split to calculate the measure. So, a buffer of 1.1x the distance from nhdFlowline to intersectionPoint was used to split the nhdFlowline.