Clickpoint On/Off flowline functioning
Reorganized the code, added several new functions. Only split catchment and current catch polygons are returned when the clickpoint is off a flowline. Merged catchment and current catchment polygons when clickpoint if on a flowline. DownstreamPath is only returned when clickpoint is off flowline. DownstreamFlowline and UpstreamFlowline are both return in both situations. There are still issues when the intersectionPoint is at the start or end of a NHD Flowline.